Add alternative NewDirect() method.
This version of the New method does not rely on the OMXPlayer to
write the DBUS address and PID to a temporary file (which is
actually performed by a wrapper script that invokes
omxplayer.bin eventually). Instead, this method relies on the
DBUS_SESSION_BUS_ADDRESS being set as an environment variable,
which is then passed-on to the newly created OMXPlayer instance.

This new instance will connect to that session bus, and so will
this library.

This makes it possible to start your application using tools like
`dbus-launch my_omxplayer_controller` in an environment where a
session bus is not available.